Implement local bookmarks, #715
We already had implementation of bookmarks which relies on the server
but most servers don't support this yet so we needed a local fallback.
This commits adds such a fallback. Bookmarked statuses are saved to
the DB and never cleaned up. We fetch them likely any other statuses
from DB with additional condition.