Ensure the "Apply" button is always visible
On smaller devices the notification filter listview may be longer than the
screen height, and pushes the "Apply" button out of sight.

Fix this by:

- Set the height of the listview to 0dp and its layout_weight to 1
- Set the layout_weight of the button to 0

This ensures the button always appears (because the listview height is 0dp)
and the listview then expands to fill any remaining space (because the
layout_weight is 1).

Fixes https://github.com/tuskyapp/Tusky/issues/2985