Revert removal of Federated timeline from defaults
Due to the removal, people have begun to stop using tusky and or finding another app to use that has a federated timeline. And adding the federated timeline back is extremely difficult and frustrating for uses.

This reverts the removal of the federated timeline, by adding it back to the defaults, resulting in a 5 button navigation bar.