Provide a preference to scale all UI text
Font scaling is applied in addition to any scaling set in Android system
preferences. So if the user set the Android font size to largest (a 1.3x
increase) and then sets the preference to 120%, the total change is 1.56x.

Create SliderPreference to adjust the preference.

- Use Slider, which supports float values and step sizes > 1
- Display the selected value in the preference's summary
- Provide buttons to increment / decrement the value

Restart the activity if the preference changes so that the user sees the
impact of the change immediately. Fix a bug in PreferencesActivity where the
"EXTRA_RESTART_ON_BACK" intent was never processed. Fix this to ensure that
other activities are restarted so the new font scale takes effect.

Implement the scaling in BaseActivity by overriding onAttachBaseContext, and
providing a wrapped context with the font scaling applied.

Fixes https://github.com/tuskyapp/Tusky/issues/2982, https://github.com/tuskyapp/Tusky/issues/2461