Support Atkinson Hyperlegible and Opendyslexic fonts
Provide two alternative fonts for use throughout the entire app.

- Atkinson Hyperlegible, from the Braille Institute, claims to improve
  readability for low-vision users.

- OpenDyslexic claims to improve readability for people with dyslexia.

Both fonts are available under a permissive license that allows bundling
with the app. Include the license information so it shows up in
LicenseActivity.

Add a preference to allow the user to switch between the fonts. The
preference previews the font in the list so the user can see the
difference.