Allow "Back" to open the nav drawer
Open the nav drawer if the user is on the first tab and presses "Back",

If the nav drawer is open, *and* it was opened by pressing "Back" then
pressing "Back" again will finish() the activity.

Otherwise, pressing "Back" with the nav drawer open will close the drawer, as
before.

Fixes https://github.com/tuskyapp/Tusky/issues/3323