Show reblog/favourite confirmations as menus not dialogs
The previous code used dialogs and displayed the text of the status when
reblogging or favouriting.

This didn't work when the post just contained images, and other material
from the status (content warning, polls) was not shown either.

Fix this by displaying a popup menu instead. The status remains visible so
the user can clearly see what they're acting on.

In addition, this lays the groundwork for supporting a long-press menu
in the future to allow the user to reblog/favourite from a different
account.

Fixes https://github.com/tuskyapp/Tusky/issues/3308