Remove NetworkTimelineRemoteMediator
The layering of this code was weird. There's paging sources and remote
mediators making calls "up" in to the viewmodel, so they had a cyclic
dependency on each other.

And there's a remote mediator when there's no local database to cache to,
which made things more complicated.

So:

1. Remove the remote mediator
2. Put the code to fetch statuses in to the paging source
3. Push the a variant of the "TimelineKind" type in to the paging source
   (probably not its final location), as this needs to be as low down the
   dependency tree as possible to avoid cycle.
   - Implement it as a sealed class instead of an enum, so that the
     subclasses can hold non-null Strings where necessary, instead of relying
     on `!!`
4. Comment out some code that either needs to be deleted or refactored, I'm
   not sure which yet.

This builds and runs, and shows network timelines (local, federated, etc).
Some functionality it lost (e.g., retaining state about whether or not the
user has expanded a status) but that will be restored in future commits.