Migrate from TimelineViewModel.Kind to TimelineKind
The old type was an enum, the new type is a hierarchy of sealed classes.

This allows each class to carry with it the information necessary for that
timeline.

For example, the PublicFederated timeline needs no additional data. But a
timeline for a hashtag does.

This simplifies TimelineFragment.newInstance(). Before the called had to
know which optional parameters to pass in, or whether they should actually
call newHashtagInstance.

Now they just construct a timeline of the appropriate kind, with the correct
data, and pass that in.

This also means that initialisation of TimelineFragment does not need to
unpack data from multiple extras in the launch intent, everything is in the
parcelized TimelineKind.