mklanguages: Manage language lists in donottranslate.xml
mklanguages processes values-*/strings.xml entries and uses that to determine
which languages can be used.

It uses the ICU database to generate the display name for each language,
and to list them in sorted order.

Add `runtools` scripts to make it easy to run this and other tools.

Update donottranslate.xml with comments that show the language code and
English name for each language.