Remove AccountEntity.activeNotifications
This was tracking (as a JSON string) an array of currently shown Android notifications. However, this could get out of sync with the actual display, and there was no mechanism to get it back in sync.

Replace it by identifying each Android notification with a tuple of `(tag = mastodon-notification-id, id = tusky-account-id)`. Use this to reliably find existing notifications in NotificationHelper.make().

Cancelling notifications for an account must iterate over all the notifications, and individually remove the notifications that exist for that account.

Adjust the summary notification:
- Ensure the summary notification and the child notifications have the same group key
- Dismiss the summary notification if there is only one child notification

NotificationClearBroadcastReceiver is no longer needed, so remove it, and the need for deletePendingIntent.