Comment out FilterResult.status_matches
it is unclear what the correct type should be.

The Mastodon documentation is inconsistent, it's either a `String?`, or a `List<String>?`.

Getting this wrong crashes during JSON parsing.

Since the field isn't used anywhere in Tusky yet it's easier to comment this out so it is not deserialized.

This can be restored when https://github.com/mastodon/mastodon/issues/24900 is fixed.