Perform long running operations with an external scope
Using the view model's scope for long running operations is a problem.
If the user navigates away and the view model is killed the operation
will be cancelled.

This is OK when fetching data. But sending data using the API (e.g.,
bookmarking a status) should continue in the background after the
user navigates away.

- Provide an ApplicationScope injectable
- Inject it to the repositories that need it
- Use the scope when performing potentially long-running operations